refactor: rename repository files for consistency and update import paths
