+++ src/repositories/index.ts
@@ -1,5 +1,5 @@
-export * from './Project';
-export * from './Tag';
-export * from './Task';
-export * from './Status';
+export * from './Project.repository';
+export * from './Tag.repository';
+export * from './Task.repository';
+export * from './Status.repository';
 export * from './User';
